Tighten security of config files and db
Sensitive and secret information can be stored in these files,
so we don't want them accessible to everyone on the machine.

Here we move all file locations into a directory in snap_common
that only root has access to.
This means:
- headscale can only be used and interacted with as root
- the sensitive files can only be read by root

This should be fine from a usability point of view,
because the daemon is already a system service running as root,
so only allowing root
to use the admin commands via headscale seems reasonable.