Add security hardening guidelines (#15)
Also move the security section to a new doc,
since it's getting too large to stay in the snapcraft description.

The scope of configuration and operation for Headscale is very large,
with many options affecting security,
so this only provides some initial general guidelines
for thinking about security in running Headscale.